requirements.txt
omegaconf==2.1.1
einops==0.3.0
#transformers==4.19.2
open_clip_torch==2.0.2
torchmetrics==0.6.0
augly==1.0.0
scikit-image==0.20.0
pytorch-fid==0.3.0
pandas==1.5.3
matplotlib

pytorch-lightning==1.4.2
scipy
transformers
diffusers
invisible-watermark
kornia
# torch==1.12.1+cu113